requirements.txt
attrs==18.2.0
CacheControl==0.12.5
cachy==0.2.0
certifi==2019.3.9
cleo==0.6.8
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Login==0.4.1
html5lib==1.0.1
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.0
lockfile==0.12.2
MarkupSafe==1.1.1
pastel==0.1.0
peewee==3.9.2
peewee-db-evolve==3.7.0
pkginfo==1.5.0.1
poetry==0.12.11
psycopg2-binary==2.7.7
PyJWT==1.7.1
pylev==1.3.0
pyrsistent==0.14.11
python-dotenv==0.10.1
requests-toolbelt==0.8.0
shellingham==1.2.8
six==1.12.0
tomlkit==0.5.3
webencodings==0.5.1
Werkzeug==0.14.1
wincertstore==0.2
